Much refactoring and effort for little gain
This commit brings many changes, mostly in the Flutter code, to make the
app slightly functional. Right now, there is a bug that causes
apparently successful CreateTopicMap commands to have no effect on the
results of GetTopicMaps queries. I'm probably missing something obvious,
but in the meantime I think it's time for a commit.