Attempt to fix db issue with sync after commit
Since one theory to explain the issue where topic maps created in one
session are not visible in future sessions is that the transactions
creating topic maps are not flushed to disk, manually sync the database
after every commit. Sync errors would show up in the log.

Unfortunately, this has not fixed the bug. However, it's not a bad idea
and it will help to continue ruling out this possibility in the future.