TopicPage is always a tabbed view of forms
This is another UI idea. The topic page could always be a tabbed view
showing the focused topic as a sibling to other topics, or even topic
map to other topic maps. Then it's easier to package the name into the
body of the page, and treat teach page like a form for editing that
topic.