Initial effort to support adding occurrences
It partly works, but occurrence value is soon lost. The code is almost
the same as for names, and largely deduplicated, but still there is
something wrong in the occurrence-specific code.

This commit also replaces the FAB on TopicPage with one that supports
Material design's "speed dial". Unfortunately, it's no longer
horizontally centered. This can be fixed later.

This commit also maks the headings on TopicPage all-caps.