Add a CLI in Go
This commit adds a third independent branch of development: the original
Flutter & Go as a prototype mobile app, a more recent JavaScript attempt
at producing a PWA with an interesting editor, and now a CLI in Go
intended to be a bare-bones proof of concept.

I'm keeping all three branches merged together like this because, well,
you never know which ones might some day combine really well. Maybe just
a .proto file from one will be useful in another. Not that this has
happened yet, but it might!

Documentation updates coming up.