notes: switch up the main interfaces
I knew I'd have to repeat work I'd already done for the removed pbdb and
genji integrations in order to make the textile integration work
properly. This is mostly about making sure I won't have to repeat much
of that work again. It also adds a bit of in-memory caching for the
ops that use a transaction, and for some reason I thought
"IsolatedReadWriteCloser" was a better name than "Database".

Anyway, it actually works! At least as well as it used to, that is, but
now it does all that with ThreadsDB, which is encrypted and peer-to-peer
replicated across untrusted nodes. It's not actually useful yet.