packages/nm_gql_go_link: new plugin
This plugin will aim to implement minimal knowledge of any particular
GraphQL schema, hopefully leading to an implementation that can be
generalized to any GraphQL schema with some clever build rules or
something like that. Until then, it will be a neat way to connect a
Flutter app to a Go implementation of a GraphQL service.